+++ src/main/java/dinkplugin/message/NotificationBody.java
@@ -1,19 +1,26 @@
 package dinkplugin.message;
 
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.Singular;
 import lombok.Value;
 
-import java.util.ArrayList;
+import java.util.LinkedList;
 import java.util.List;
 
 @Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
 public class NotificationBody<T> {
     private NotificationType type;
     private T extra;
-    private String playerName;
-
     private String content;
-    private List<Embed> embeds = new ArrayList<>();
+    private String playerName;
+    @Singular
+    private List<Embed> embeds = new LinkedList<>();
 
     @Value
     public static class Embed {