+++ src/main/java/dinkplugin/notifiers/ClueNotifier.java
@@ -1,30 +1,32 @@
 package dinkplugin.notifiers;
 
 import dinkplugin.DinkPlugin;
+import dinkplugin.DinkPluginConfig;
 import dinkplugin.message.NotificationBody;
 import dinkplugin.message.NotificationType;
 import dinkplugin.Utils;
 import dinkplugin.notifiers.data.ClueNotificationData;
 import dinkplugin.notifiers.data.SerializedItemStack;
-import lombok.Getter;
 import net.runelite.api.ItemComposition;
 import net.runelite.api.events.WidgetLoaded;
 import net.runelite.api.widgets.Widget;
 import net.runelite.api.widgets.WidgetID;
 import net.runelite.api.widgets.WidgetInfo;
 import net.runelite.client.util.QuantityFormatter;
+import org.apache.commons.lang3.StringUtils;
 
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.atomic.AtomicLong;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public class ClueNotifier extends BaseNotifier {
     private static final Pattern CLUE_SCROLL_REGEX = Pattern.compile("You have completed (?<scrollCount>\\d+) (?<scrollType>\\w+) Treasure Trails\\.");
 
-    @Getter
     private final Map<Integer, Integer> clueItems = new HashMap<>();
     private boolean clueCompleted = false;
     private String clueCount = "";
@@ -34,19 +36,21 @@ public class ClueNotifier extends BaseNotifier {
         super(plugin);
     }
 
+    @Override
+    public boolean isEnabled() {
+        return plugin.getConfig().notifyClue() && super.isEnabled();
+    }
+
     public void onChatMessage(String chatMessage) {
-        if (plugin.getConfig().notifyClue()) {
+        if (isEnabled()) {
             Matcher clueMatcher = CLUE_SCROLL_REGEX.matcher(chatMessage);
             if (clueMatcher.find()) {
-                String numberCompleted = clueMatcher.group("scrollCount");
-                String scrollType = clueMatcher.group("scrollType");
+                clueCount = clueMatcher.group("scrollCount");
+                clueType = clueMatcher.group("scrollType");
 
                 if (clueCompleted) {
-                    this.handleNotify(numberCompleted, scrollType);
-                    clueCompleted = false;
+                    this.handleNotify();
                 } else {
-                    clueType = scrollType;
-                    clueCount = numberCompleted;
                     clueCompleted = true;
                 }
             }
@@ -54,32 +58,25 @@ public class ClueNotifier extends BaseNotifier {
     }
 
     public void onWidgetLoaded(WidgetLoaded event) {
-        if (event.getGroupId() == WidgetID.CLUE_SCROLL_REWARD_GROUP_ID) {
+        if (event.getGroupId() == WidgetID.CLUE_SCROLL_REWARD_GROUP_ID && isEnabled()) {
             Widget clue = plugin.getClient().getWidget(WidgetInfo.CLUE_SCROLL_REWARD_ITEM_CONTAINER);
             if (clue != null) {
                 clueItems.clear();
                 Widget[] children = clue.getChildren();
-
-                if (children == null) {
-                    return;
-                }
+                if (children == null) return;
 
                 for (Widget child : children) {
-                    if (child == null) {
-                        continue;
-                    }
+                    if (child == null) continue;
 
                     int quantity = child.getItemQuantity();
                     int itemId = child.getItemId();
-
                     if (itemId > -1 && quantity > 0) {
-                        clueItems.put(itemId, quantity);
+                        clueItems.merge(itemId, quantity, Integer::sum);
                     }
                 }
 
                 if (clueCompleted) {
-                    this.handleNotify(clueCount, clueType);
-                    clueCompleted = false;
+                    this.handleNotify();
                 } else {
                     clueCompleted = true;
                 }
@@ -87,55 +84,51 @@ public class ClueNotifier extends BaseNotifier {
         }
     }
 
-    private void handleNotify(String numberCompleted, String clueType) {
-        if (plugin.isIgnoredWorld()) return;
-        NotificationBody<ClueNotificationData> messageBody = new NotificationBody<>();
-
+    private void handleNotify() {
         StringBuilder lootMessage = new StringBuilder();
-        long totalPrice = 0;
-        List<SerializedItemStack> itemStacks = new ArrayList<>();
+        AtomicLong totalPrice = new AtomicLong();
+        List<SerializedItemStack> itemStacks = new ArrayList<>(clueItems.size());
+        List<NotificationBody.Embed> embeds = new ArrayList<>(plugin.getConfig().clueShowItems() ? clueItems.size() : 0);
 
-        for (Integer itemId : clueItems.keySet()) {
-            if (lootMessage.length() > 0) {
-                lootMessage.append("\n");
-            }
-            int quantity = clueItems.get(itemId);
+        clueItems.forEach((itemId, quantity) -> {
+            if (lootMessage.length() > 0) lootMessage.append('\n');
+
             int price = plugin.getItemManager().getItemPrice(itemId);
-            totalPrice += (long) price * quantity;
-            lootMessage.append(getItem(itemId, clueItems.get(itemId), messageBody));
-
             ItemComposition itemComposition = plugin.getItemManager().getItemComposition(itemId);
-            itemStacks.add(new SerializedItemStack(itemId, quantity, price, itemComposition.getName()));
+            SerializedItemStack stack = new SerializedItemStack(itemId, quantity, price, itemComposition.getName());
+
+            totalPrice.addAndGet(stack.getTotalPrice());
+            itemStacks.add(stack);
+            lootMessage.append(getItemMessage(stack, embeds));
+        });
+
+        if (totalPrice.get() >= plugin.getConfig().clueMinValue()) {
+            String notifyMessage = StringUtils.replaceEach(
+                plugin.getConfig().clueNotifyMessage(),
+                new String[] { "%USERNAME%", "%CLUE%", "%COUNT%", "%TOTAL_VALUE%", "%LOOT%" },
+                new String[] { Utils.getPlayerName(plugin.getClient()), clueType, clueCount, QuantityFormatter.quantityToStackSize(totalPrice.get()), lootMessage.toString() }
+            );
+            createMessage(DinkPluginConfig::clueSendImage, NotificationBody.builder()
+                .content(notifyMessage)
+                .extra(new ClueNotificationData(clueType, Integer.parseInt(clueCount), itemStacks))
+                .type(NotificationType.CLUE)
+                .embeds(embeds)
+                .build());
         }
 
-        if (totalPrice < plugin.getConfig().clueMinValue()) {
-            return;
-        }
-
-        String notifyMessage = plugin.getConfig().clueNotifyMessage()
-            .replaceAll("%USERNAME%", Utils.getPlayerName())
-            .replaceAll("%CLUE%", clueType)
-            .replaceAll("%COUNT%", numberCompleted)
-            .replaceAll("%TOTAL_VALUE%", QuantityFormatter.quantityToStackSize(totalPrice))
-            .replaceAll("%LOOT%", lootMessage.toString());
-        messageBody.setContent(notifyMessage);
-        ClueNotificationData extra = new ClueNotificationData();
-        extra.setClueType(clueType);
-        extra.setNumberCompleted(Integer.parseInt(numberCompleted));
-        extra.setItems(itemStacks);
-        messageBody.setExtra(extra);
-        messageBody.setType(NotificationType.CLUE);
-        messageHandler.createMessage(plugin.getConfig().clueSendImage(), messageBody);
+        this.reset();
     }
 
-    private String getItem(int itemId, int quantity, NotificationBody<ClueNotificationData> messageBody) {
-        int price = plugin.getItemManager().getItemPrice(itemId);
-        long totalPrice = (long) price * quantity;
-        ItemComposition itemComposition = plugin.getItemManager().getItemComposition(itemId);
+    private String getItemMessage(SerializedItemStack item, Collection<NotificationBody.Embed> embeds) {
+        if (plugin.getConfig().clueShowItems())
+            embeds.add(new NotificationBody.Embed(new NotificationBody.UrlEmbed(Utils.getItemImageUrl(item.getId()))));
+        return String.format("%s x %s (%s)", item.getQuantity(), item.getName(), QuantityFormatter.quantityToStackSize(item.getTotalPrice()));
+    }
 
-        if (plugin.getConfig().clueShowItems()) {
-            messageBody.getEmbeds().add(new NotificationBody.Embed(new NotificationBody.UrlEmbed(Utils.getItemImageUrl(itemId))));
-        }
-        return String.format("%s x %s (%s)", quantity, itemComposition.getName(), QuantityFormatter.quantityToStackSize(totalPrice));
+    private void reset() {
+        this.clueCompleted = false;
+        this.clueCount = "";
+        this.clueType = "";
+        this.clueItems.clear();
     }
 }