+++ src/main/java/dinkplugin/notifiers/CollectionNotifier.java
@@ -1,10 +1,12 @@
 package dinkplugin.notifiers;
 
 import dinkplugin.DinkPlugin;
+import dinkplugin.DinkPluginConfig;
 import dinkplugin.message.NotificationBody;
 import dinkplugin.message.NotificationType;
 import dinkplugin.Utils;
 import dinkplugin.notifiers.data.CollectionNotificationData;
+import org.apache.commons.lang3.StringUtils;
 
 import javax.inject.Inject;
 import java.util.regex.Matcher;
@@ -18,8 +20,13 @@ public class CollectionNotifier extends BaseNotifier {
         super(plugin);
     }
 
+    @Override
+    public boolean isEnabled() {
+        return plugin.getConfig().notifyCollectionLog() && super.isEnabled();
+    }
+
     public void onChatMessage(String chatMessage) {
-        if (!plugin.getConfig().notifyCollectionLog()) return;
+        if (!isEnabled()) return;
 
         Matcher collectionMatcher = COLLECTION_LOG_REGEX.matcher(chatMessage);
         if (collectionMatcher.find()) {
@@ -28,16 +35,16 @@ public class CollectionNotifier extends BaseNotifier {
     }
 
     private void handleNotify(String itemName) {
-        if (plugin.isIgnoredWorld()) return;
-        String notifyMessage = plugin.getConfig().collectionNotifyMessage()
-            .replaceAll("%USERNAME%", Utils.getPlayerName())
-            .replaceAll("%ITEM%", itemName);
-        NotificationBody<CollectionNotificationData> b = new NotificationBody<>();
-        b.setContent(notifyMessage);
-        CollectionNotificationData extra = new CollectionNotificationData();
-        extra.setItemName(itemName);
-        b.setExtra(extra);
-        b.setType(NotificationType.COLLECTION);
-        messageHandler.createMessage(plugin.getConfig().collectionSendImage(), b);
+        String notifyMessage = StringUtils.replaceEach(
+            plugin.getConfig().collectionNotifyMessage(),
+            new String[] { "%USERNAME%", "%ITEM%" },
+            new String[] { Utils.getPlayerName(plugin.getClient()), itemName }
+        );
+
+        createMessage(DinkPluginConfig::collectionSendImage, NotificationBody.builder()
+            .content(notifyMessage)
+            .extra(new CollectionNotificationData(itemName))
+            .type(NotificationType.COLLECTION)
+            .build());
     }
 }