+++ src/main/java/dinkplugin/notifiers/QuestNotifier.java
@@ -1,6 +1,7 @@
 package dinkplugin.notifiers;
 
 import dinkplugin.DinkPlugin;
+import dinkplugin.DinkPluginConfig;
 import dinkplugin.message.NotificationBody;
 import dinkplugin.message.NotificationType;
 import dinkplugin.Utils;
@@ -8,6 +9,7 @@ import dinkplugin.notifiers.data.QuestNotificationData;
 import net.runelite.api.events.WidgetLoaded;
 import net.runelite.api.widgets.Widget;
 import net.runelite.api.widgets.WidgetInfo;
+import org.apache.commons.lang3.StringUtils;
 
 import static net.runelite.api.widgets.WidgetID.QUEST_COMPLETED_GROUP_ID;
 
@@ -17,27 +19,32 @@ public class QuestNotifier extends BaseNotifier {
         super(plugin);
     }
 
+    @Override
+    public boolean isEnabled() {
+        return plugin.getConfig().notifyQuest() && super.isEnabled();
+    }
+
     public void onWidgetLoaded(WidgetLoaded event) {
-        if (event.getGroupId() == QUEST_COMPLETED_GROUP_ID && plugin.getConfig().notifyQuest()) {
+        if (event.getGroupId() == QUEST_COMPLETED_GROUP_ID && isEnabled()) {
             Widget quest = plugin.getClient().getWidget(WidgetInfo.QUEST_COMPLETED_NAME_TEXT);
             if (quest != null) {
-                String questWidget = quest.getText();
-                this.handleNotify(questWidget);
+                this.handleNotify(quest.getText());
             }
         }
     }
 
     private void handleNotify(String questText) {
-        if (plugin.isIgnoredWorld()) return;
-        String notifyMessage = plugin.getConfig().questNotifyMessage()
-            .replaceAll("%USERNAME%", Utils.getPlayerName())
-            .replaceAll("%QUEST%", Utils.parseQuestWidget(questText));
-        NotificationBody<QuestNotificationData> body = new NotificationBody<>();
-        body.setContent(notifyMessage);
-        QuestNotificationData extra = new QuestNotificationData();
-        extra.setQuestName(Utils.parseQuestWidget(questText));
-        body.setExtra(extra);
-        body.setType(NotificationType.QUEST);
-        messageHandler.createMessage(plugin.getConfig().questSendImage(), body);
+        String parsed = Utils.parseQuestWidget(questText);
+        String notifyMessage = StringUtils.replaceEach(
+            plugin.getConfig().questNotifyMessage(),
+            new String[] { "%USERNAME%", "%QUEST%" },
+            new String[] { Utils.getPlayerName(plugin.getClient()), parsed }
+        );
+
+        createMessage(DinkPluginConfig::questSendImage, NotificationBody.builder()
+            .content(notifyMessage)
+            .extra(new QuestNotificationData(parsed))
+            .type(NotificationType.QUEST)
+            .build());
     }
 }