+++ src/main/java/dinkplugin/notifiers/SpeedrunNotifier.java
@@ -1,6 +1,7 @@
 package dinkplugin.notifiers;
 
 import dinkplugin.DinkPlugin;
+import dinkplugin.DinkPluginConfig;
 import dinkplugin.message.NotificationBody;
 import dinkplugin.message.NotificationType;
 import dinkplugin.Utils;
@@ -9,6 +10,7 @@ import lombok.extern.slf4j.Slf4j;
 import net.runelite.api.Client;
 import net.runelite.api.events.WidgetLoaded;
 import net.runelite.api.widgets.Widget;
+import org.apache.commons.lang3.StringUtils;
 
 import java.time.Duration;
 import java.util.regex.Matcher;
@@ -16,6 +18,7 @@ import java.util.regex.Pattern;
 
 @Slf4j
 public class SpeedrunNotifier extends BaseNotifier {
+    private static final Pattern TIME_PATTERN = Pattern.compile("(?<minutes>\\d+):(?<seconds>\\d{2})\\.(?<fractional>\\d{2})");
     private static final int SPEEDRUN_COMPLETED_GROUP_ID = 781;
     private static final int SPEEDRUN_COMPLETED_QUEST_NAME_CHILD_ID = 4;
     private static final int SPEEDRUN_COMPLETED_DURATION_CHILD_ID = 10;
@@ -25,8 +28,13 @@ public class SpeedrunNotifier extends BaseNotifier {
         super(plugin);
     }
 
+    @Override
+    public boolean isEnabled() {
+        return plugin.getConfig().notifySpeedrun(); // intentionally doesn't call super
+    }
+
     public void onWidgetLoaded(WidgetLoaded event) {
-        if (event.getGroupId() == SPEEDRUN_COMPLETED_GROUP_ID && plugin.getConfig().notifySpeedrun()) {
+        if (event.getGroupId() == SPEEDRUN_COMPLETED_GROUP_ID && isEnabled()) {
             Client client = plugin.getClient();
             Widget questName = client.getWidget(SPEEDRUN_COMPLETED_GROUP_ID, SPEEDRUN_COMPLETED_QUEST_NAME_CHILD_ID);
             Widget duration = client.getWidget(SPEEDRUN_COMPLETED_GROUP_ID, SPEEDRUN_COMPLETED_DURATION_CHILD_ID);
@@ -40,37 +48,28 @@ public class SpeedrunNotifier extends BaseNotifier {
     }
 
     private void attemptNotify(String questName, String duration, String pb) {
-        Duration bestTime = this.parseTime(pb);
-        Duration currentTime = this.parseTime(duration);
+        Duration bestTime = parseTime(pb);
+        Duration currentTime = parseTime(duration);
         boolean isPb = bestTime.compareTo(currentTime) >= 0;
         if (!isPb && plugin.getConfig().speedrunPBOnly()) {
             return;
         }
-        // pb or notifying on non-pb; take the right string and format placeholders
-        String notifyMessage = plugin.getConfig().speedrunMessage();
-        if (isPb) {
-            notifyMessage = plugin.getConfig().speedrunPBMessage();
-        }
 
-        notifyMessage = notifyMessage
-            .replaceAll("%USERNAME%", Utils.getPlayerName())
-            .replaceAll("%QUEST%", questName)
-            .replaceAll("%TIME%", duration)
-            .replaceAll("%BEST%", pb);
-        NotificationBody<SpeedrunNotificationData> body = new NotificationBody<>();
-        body.setContent(notifyMessage);
-        SpeedrunNotificationData extra = new SpeedrunNotificationData();
-        extra.setQuestName(questName);
-        extra.setPersonalBest(bestTime.toString());
-        extra.setCurrentTime(currentTime.toString());
-        body.setExtra(extra);
-        body.setType(NotificationType.SPEEDRUN);
-        messageHandler.createMessage(plugin.getConfig().speedrunSendImage(), body);
+        // pb or notifying on non-pb; take the right string and format placeholders
+        String notifyMessage = StringUtils.replaceEach(
+            isPb ? plugin.getConfig().speedrunPBMessage() : plugin.getConfig().speedrunMessage(),
+            new String[] { "%USERNAME%", "%QUEST%", "%TIME%", "%BEST%" },
+            new String[] { Utils.getPlayerName(plugin.getClient()), questName, duration, pb }
+        );
+
+        createMessage(DinkPluginConfig::speedrunSendImage, NotificationBody.builder()
+            .content(notifyMessage)
+            .extra(new SpeedrunNotificationData(questName, bestTime.toString(), currentTime.toString()))
+            .type(NotificationType.SPEEDRUN)
+            .build());
     }
 
-    private static final Pattern TIME_PATTERN = Pattern.compile("(?<minutes>\\d+):(?<seconds>\\d{2})\\.(?<fractional>\\d{2})");
-
-    public Duration parseTime(String in) {
+    private static Duration parseTime(String in) {
         Matcher m = TIME_PATTERN.matcher(in);
         if (!m.find()) return Duration.ofMillis(0);
         int minutes = Integer.parseInt(m.group("minutes"));